START adding sfneal/builders test suite
