+++ tests/BuilderTestCase.php
@@ -1,33 +1,16 @@
 <?php
 
-namespace Sfneal\Builders\Tests;
+namespace Sfneal\Models\Tests;
 
 use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Orchestra\Testbench\TestCase as OrchestraTestCase;
-use Sfneal\Builders\Tests\Models\People;
-use Sfneal\Builders\Tests\Providers\TestingServiceProvider;
+use Sfneal\Models\Tests\Models\People;
 
-class TestCase extends OrchestraTestCase
+class BuilderTestCase extends TestCase
 {
-    use RefreshDatabase;
-
     /**
      * @var People|Collection
      */
     public $models;
-
-    protected function getPackageProviders($app)
-    {
-        return TestingServiceProvider::class;
-    }
-
-    protected function getEnvironmentSetUp($app)
-    {
-        include_once __DIR__.'/migrations/create_people_table.php.stub';
-
-        (new \CreatePeopleTable())->up();
-    }
 
     /**
      * Setup the test environment.
