+++ tests/Models/People.php
@@ -2,13 +2,22 @@
 
 namespace Sfneal\Models\Tests\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Sfneal\Models\Model;
+use Illuminate\Database\Eloquent\Model;
+use Sfneal\Builders\QueryBuilder;
 use Sfneal\Models\Tests\Factories\PeopleFactory;
 
 class People extends Model
 {
     use HasFactory;
+
+    /**
+     * The attributes that should be mutated to dates.
+     *
+     * @var array
+     */
+    protected $dates = ['deleted_at', 'updated_at', 'created_at'];
 
     protected $table = 'people';
     protected $primaryKey = 'person_id';
@@ -25,9 +34,35 @@ class People extends Model
         'zip',
     ];
 
+    /**
+     * Model Factory.
+     *
+     * @return PeopleFactory
+     */
     protected static function newFactory(): PeopleFactory
     {
         return new PeopleFactory();
+    }
+
+    /**
+     * Query Builder.
+     *
+     * @param $query
+     * @return QueryBuilder
+     */
+    public function newEloquentBuilder($query)
+    {
+        return new QueryBuilder($query);
+    }
+
+    /**
+     * Query Builder method for improved type hinting.
+     *
+     * @return QueryBuilder|Builder
+     */
+    public static function query()
+    {
+        return parent::query();
     }
 
     public function getNameFullAttribute(): string
@@ -43,5 +78,10 @@ class People extends Model
     public function getAddressFullAttribute(): string
     {
         return "{$this->address}, {$this->city}, {$this->state} {$this->zip}";
+    }
+
+    public function getAgeAttribute($value): int
+    {
+        return intval($value);
     }
 }
