+++ tests/Providers/TestingServiceProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\Builders\Tests\Providers;
+namespace Sfneal\Models\Tests\Providers;
 
 use Illuminate\Support\ServiceProvider;
 
@@ -11,7 +11,7 @@ class TestingServiceProvider extends ServiceProvider
         // Publish migration file (if not already published)
         if (! class_exists('CreatePeopleTable')) {
             $this->publishes([
-                __DIR__.'/../migrations/create_people_table.php.stub' => database_path(
+                __DIR__ . '/../migrations/create_people_table.php.stub' => database_path(
                     'migrations/'.date('Y_m_d_His', time()).'_create_people_table.php'
                 ),
             ], 'migration');
