+++ tests/TestCase.php
@@ -1,26 +1,25 @@
 <?php
 
+
 namespace Sfneal\Models\Tests;
 
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
-use Sfneal\Models\Tests\Models\People;
+use Sfneal\Models\Tests\Providers\TestingServiceProvider;
 
 class TestCase extends OrchestraTestCase
 {
-    /**
-     * @var People
-     */
-    public $model;
+    use RefreshDatabase;
 
-    /**
-     * Setup the test environment.
-     *
-     * @return void
-     */
-    public function setUp(): void
+    protected function getPackageProviders($app)
     {
-        parent::setUp();
+        return TestingServiceProvider::class;
+    }
 
-        $this->model = People::factory()->make();
+    protected function getEnvironmentSetUp($app)
+    {
+        include_once __DIR__ . '/migrations/create_people_table.php.stub';
+
+        (new \CreatePeopleTable())->up();
     }
 }
